refactor(account): 目录结构统一 + 颜色走主题
- main.py → page.py，与 editor/history 命名一致
- account_form.py 从 dialogs/ 提到同级，删除多余目录
- __init__.py 导出 AccountDialog，main_window 改为从包导入
- 按钮图标去掉硬编码 #434343，走主题默认色
- QSS 状态文字改用 {success}/{danger} 主题变量
