+++ src/danmaku_sender/ui/account_manager/__init__.py
@@ -1,0 +1,3 @@
+from .page import AccountDialog
+
+__all__ = ["AccountDialog"]
